Fix bug in the application of Bland's rule
I misunderstood how this works. The score of the candidate pivot does
not need to be positive. Only the pivot cell needs to be positive, it
seems.

Also needed to fix extraction of basic solutions. It is not enough to
look for a 0 in the objective row: we must also verify that the variable
is actually basic.

I'm still not entirely sure how non-unique solutions are detected,
either.